RED: Semantic validator refactoring
Introduced a cache infrastructure for merchant name and GUID in the
common package as a Spring bean.

The SemanticValidatorFactory, now in the common package, is a Spring
bean so that the cache can be injected into it.

Spring injects a SemanticValidatorFactory into the ApplicationService,
which uses it to get the SemanticValidatorUseCase for the
CreateMerchantService.create method.