Handling multiple issues on the same response
This commit provides a solution for gathering the semantic issues in
order to return all of them in one go to the user. The solution
implemented is not ideal, as it has the drawnback of having to pass the
collection to the SemanticValidator::validate method.

It also has some unit test refactoring for better structuring the
semantic validation unit tests so they are less bridle. The
CreateSubMerchantServiceTest only checks if the sub-merchant's semantic
validation method is called and ensures that its semantic validation is
running. It's no longer job of this test class to test the exceptions.

Finally, it changes some unit tests SemanticValidatorFactoryTest class
refactoring to get rid of Mockito.