Merchant generalization and syntactic validator
This commit gets rid of the merchant generalization. Now, only
sub-merchants are created. This means that a merchant type is no longer
required, as the class type determines the type of the merchant before
persisting it to the DB (later on).

This commit also changes where the syntactic validator is being called.
All the domain entity Jakarta Bean Validation annotations have been
removed. The validations are now restricted to the command entities. The
validations are now kicked in from the ApplicationService class.