Method renaiming
In the CreateSubMerchantUseCase interface, the method has been renamed
to executeCommand. The rationale is to have a method name shared across
all the use cases (in the future) so that they can be overridden in the
ApplicationService class. This also simplifies in the adapter as it
doesn't need to know which method it needs to call from the
ApplicationService (just call the executeCommand method passing in the
command class and the ApplicationService will execute the right
overridden method).